[PCM-847] Refactor logging out into a separate module, add initial code for initialization in anonymous mode
