+++ src/client-private.js
@@ -2,44 +2,40 @@
 // These are all invoked bound to an instance of the class, but are not exposed on the
 // instance object.
 
-const request = require('superagent');
 const StatsGatherer = require('webrtc-stats-gatherer');
 const StreamingClient = require('purecloud-streaming-client');
 
 const PC_AUDIO_EL_CLASS = '__pc-webrtc-inbound';
 
-function buildUri (path, version = 'v2') {
-  path = path.replace(/^\/+|\/+$/g, ''); // trim leading/trailing /
-  return `https://api.${this._environment}/api/${version}/${path}`;
-}
-
-function requestApi (path, { method, data, version, contentType } = {}) {
-  let response = request[method || 'get'](buildUri.call(this, path, version))
-    .set('Authorization', `Bearer ${this._accessToken}`)
-    .type(contentType || 'json');
-
-  return response.send(data); // trigger request
-}
-
-function rejectErr (message, details) {
-  const error = new Error(message);
-  error.details = details;
-  this.emit('error', message, details);
-  throw error;
-}
+const { log } = require('./logging');
 
 function setupStreamingClient () {
+  if (this._streamingConnection) {
+    this.logger.warn('Existing streaming connection detected. Disconnecting and creating a new connection.');
+    this._streamingConnection.disconnect();
+  }
+
   const connectionOptions = {
     signalIceConnected: true,
-    jid: this._personDetails.chat.jabberId,
     iceTransportPolicy: this._iceTransportPolicy,
     host: this._wsHost || `wss://streaming.${this._environment}`,
     apiHost: this._environment,
-    logger: this.logger,
-    authToken: this._accessToken
+    logger: this.logger
   };
 
-  this._log('debug', 'Streaming client WebSocket connection options', connectionOptions);
+  if (this._personDetails) {
+    connectionOptions.jid = this._personDetails.chat.jabberId;
+  }
+
+  if (this._accessToken) {
+    connectionOptions.authToken = this._accessToken;
+  }
+
+  if (this._jwt) {
+    connectionOptions.jwt = this._jwt;
+  }
+
+  log.call(this, 'debug', 'Streaming client WebSocket connection options', connectionOptions);
   this._hasConnected = false;
 
   const connection = new StreamingClient(connectionOptions);
@@ -47,14 +43,14 @@ function setupStreamingClient () {
   return connection.connect()
     .then(() => {
       this.emit('connected', { reconnect: this._hasConnected });
-      this._log('info', 'PureCloud streaming client connected', { reconnect: this._hasConnected });
+      log.call(this, 'info', 'PureCloud streaming client connected', { reconnect: this._hasConnected });
       this._hasConnected = true;
       // refresh turn servers every 6 hours
       this._refreshTurnServersInterval = setInterval(this._refreshTurnServers.bind(this), 6 * 60 * 60 * 1000);
       return connection.webrtcSessions.refreshIceServers();
     })
     .then(e => {
-      this._log('info', 'PureCloud streaming client ready for WebRTC calls');
+      log.call(this, 'info', 'PureCloud streaming client ready for WebRTC calls');
     });
 }
 
@@ -90,7 +86,7 @@ function attachMedia (stream) {
 
 function onSession (session) {
   try {
-    this._log('info', 'onSession', { id: session.sid });
+    log.call(this, 'info', 'onSession', { id: session.sid });
   } catch (e) {
     // don't let log errors ruin a session
   }
@@ -108,21 +104,21 @@ function onSession (session) {
   });
   session._statsGatherer.on('stats', (data) => {
     data.conversationId = session.conversationId;
-    this._log('info', 'session:stats', data);
+    log.call(this, 'info', 'session:stats', data);
   });
   session._statsGatherer.on('traces', (data) => {
     data.conversationId = session.conversationId;
-    this._log('warn', 'session:trace', data);
+    log.call(this, 'warn', 'session:trace', data);
   });
   session.on('change:active', (session, active) => {
     if (active) {
       session._statsGatherer.collectInitialConnectionStats();
     }
-    this._log('info', 'change:active', { active, conversationId: session.conversationId, sid: session.sid });
+    log.call(this, 'info', 'change:active', { active, conversationId: session.conversationId, sid: session.sid });
   });
 
   startMedia.call(this).then(stream => {
-    this._log('debug', 'onMediaStarted');
+    log.call(this, 'debug', 'onMediaStarted');
     session.addStream(stream);
     session._outboundStream = stream;
 
@@ -139,7 +135,7 @@ function onSession (session) {
     }
 
     session.on('terminated', (session, reason) => {
-      this._log('info', 'onSessionTerminated', { conversationId: session.conversationId, reason });
+      log.call(this, 'info', 'onSessionTerminated', { conversationId: session.conversationId, reason });
       if (session._outboundStream) {
         session._outboundStream.getTracks().forEach(t => t.stop());
       }
@@ -150,7 +146,7 @@ function onSession (session) {
 }
 
 function onPendingSession (sessionInfo) {
-  this._log('info', 'onPendingSession', sessionInfo);
+  log.call(this, 'info', 'onPendingSession', sessionInfo);
   const sessionEvent = {
     id: sessionInfo.sessionId,
     autoAnswer: sessionInfo.autoAnswer,
@@ -182,9 +178,6 @@ function proxyStreamingClientEvents () {
 }
 
 module.exports = {
-  buildUri,
-  requestApi,
-  rejectErr,
   setupStreamingClient,
   proxyStreamingClientEvents
 };
