+++ test/test-utils.js
@@ -1,0 +1,195 @@
+const WebSocket = require('ws');
+const nock = require('nock');
+const sinon = require('sinon');
+const PureCloudWebrtcSdk = require('../src/client');
+
+let wss;
+let ws;
+
+function random () {
+  return `${Math.random()}`.split('.')[1];
+}
+
+function timeout (n) {
+  return new Promise(resolve => setTimeout(resolve, n));
+}
+
+const USER_ID = random();
+const PARTICIPANT_ID = random();
+
+const MOCK_USER = {
+  id: USER_ID,
+  chat: { jabberId: 'hubert.j.farnsworth@planetexpress.mypurecloud.com' }
+};
+
+const MOCK_ORG = {
+  thirdPartyOrgId: '3000'
+};
+
+const MOCK_CONVERSATION = {
+  participants: [
+    {
+      id: PARTICIPANT_ID,
+      user: {
+        id: USER_ID
+      }
+    },
+    {
+      id: random()
+    }
+  ]
+};
+
+function mockApis ({ failOrg, failUser, failStreaming, failLogs, failLogsPayload, withMedia, conversationId, participantId, withLogs } = {}) {
+  nock.cleanAll();
+  const api = nock('https://api.mypurecloud.com');
+
+  // easy to debug nock
+  // api.log(console.error);
+
+  let getOrg;
+  if (failOrg) {
+    getOrg = api.get('/api/v2/organizations/me').reply(401);
+  } else {
+    getOrg = api.get('/api/v2/organizations/me').reply(200, MOCK_ORG);
+  }
+
+  let getUser;
+  if (failUser) {
+    getUser = api.get('/api/v2/users/me').reply(401);
+  } else {
+    getUser = api.get('/api/v2/users/me').reply(200, MOCK_USER);
+  }
+
+  const getChannel = api
+    .post('/api/v2/notifications/channels?connectionType=streaming')
+    .reply(200, { id: 'somechannelid' });
+
+  const conversationsApi = nock('https://api.mypurecloud.com');
+  let getConversation;
+  if (conversationId) {
+    getConversation = conversationsApi
+      .get(`/api/v2/conversations/calls/${conversationId}`)
+      .reply(200, MOCK_CONVERSATION);
+  }
+
+  let patchConversation;
+  if (conversationId && participantId) {
+    patchConversation = conversationsApi
+      .patch(`/api/v2/conversations/calls/${conversationId}/participants/${participantId}`)
+      .reply(202, {});
+  }
+
+  global.window = global.window || {};
+  if (withMedia) {
+    window.navigator = window.navigator || {};
+    window.navigator.mediaDevices = window.navigator.mediaDevices || {};
+    window.navigator.mediaDevices.getUserMedia = () => Promise.resolve(withMedia);
+  }
+
+  global.document = {
+    createElement: sinon.stub().returns({
+      addEventListener: (evt, callback) => setTimeout(callback, 10),
+      classList: { add () {} }
+    }),
+    querySelector () {},
+    body: {
+      append () {}
+    },
+    head: {
+      appendChild: sinon.stub().callsFake((script) => {
+        global.window = global.window || {};
+      })
+    }
+  };
+
+  const sdk = new PureCloudWebrtcSdk({
+    accessToken: '1234',
+    wsHost: failStreaming ? null : 'ws://localhost:1234',
+    logger: { debug () {}, log () {}, info () {}, warn () {}, error () {} }
+    // logger: { debug () {}, log () {}, info () {}, warn: console.warn.bind(console), error: console.error.bind(console) }
+  });
+
+  let sendLogs;
+  if (withLogs) {
+    const logsApi = nock('https://api.mypurecloud.com').persist();
+
+    if (failLogsPayload) {
+      sendLogs = logsApi.post('/api/v2/diagnostics/trace').replyWithError({status: 413, message: 'test fail'});
+    } else if (failLogs) {
+      sendLogs = logsApi.post('/api/v2/diagnostics/trace').replyWithError({status: 419, message: 'test fail'});
+    } else {
+      sendLogs = logsApi.post('/api/v2/diagnostics/trace').reply(200);
+    }
+  } else {
+    sdk._optOutOfTelemetry = true;
+  }
+
+  if (wss) {
+    wss.close();
+    wss = null;
+  }
+
+  wss = new WebSocket.Server({
+    port: 1234
+  });
+  let websocket;
+  wss.on('connection', function (ws) {
+    websocket = ws;
+    ws.on('message', function (msg) {
+      // console.error('⬆️', msg);
+      const send = function (r) {
+        // console.error('⬇️', r);
+        setTimeout(function () { ws.send(r); }, 15);
+      };
+      if (msg.indexOf('<open') === 0) {
+        send('<open xmlns="urn:ietf:params:xml:ns:xmpp-framing" xmlns:stream="http://etherx.jabber.org/streams" version="1.0" from="hawk" id="d6f681a3-358c-49df-819f-b231adb3cb97" xml:lang="en"></open>');
+        if (ws.__authSent) {
+          send(`<stream:features xmlns:stream="http://etherx.jabber.org/streams"><bind xmlns="urn:ietf:params:xml:ns:xmpp-bind"></bind><session xmlns="urn:ietf:params:xml:ns:xmpp-session"></session></stream:features>`);
+        } else {
+          send('<stream:features xmlns:stream="http://etherx.jabber.org/streams"><mechanisms xmlns="urn:ietf:params:xml:ns:xmpp-sasl"><mechanism>PLAIN</mechanism></mechanisms></stream:features>');
+        }
+      } else if (msg.indexOf('<auth') === 0) {
+        if (failStreaming) {
+          send('<failure xmlns="urn:ietf:params:xml:ns:xmpp-sasl"></failure>');
+        } else {
+          send('<success xmlns="urn:ietf:params:xml:ns:xmpp-sasl"></success>');
+        }
+        ws.__authSent = true;
+      } else if (msg.indexOf('<bind') !== -1) {
+        const idRegexp = /id="(.*?)"/;
+        const id = idRegexp.exec(msg)[1];
+        send(`<iq xmlns="jabber:client" id="${id}" to="${MOCK_USER.chat.jabberId}" type="result"><bind xmlns="urn:ietf:params:xml:ns:xmpp-bind"><jid>${MOCK_USER.chat.jabberId}/d6f681a3-358c-49df-819f-b231adb3cb97</jid></bind></iq>`);
+      } else if (msg.indexOf('<session') !== -1) {
+        const idRegexp = /id="(.*?)"/;
+        const id = idRegexp.exec(msg)[1];
+        send(`<iq xmlns="jabber:client" id="${id}" to="${MOCK_USER.chat.jabberId}/d6f681a3-358c-49df-819f-b231adb3cb97" type="result"></iq>`);
+      } else if (msg.indexOf('ping') !== -1) {
+        const idRegexp = /id="(.*?)"/;
+        const id = idRegexp.exec(msg)[1];
+        send(`<iq xmlns="jabber:client" to="${MOCK_USER.chat.jabberId}" type="result" id="${id}"></iq>`);
+      } else if (msg.indexOf('extdisco') !== -1) {
+        const idRegexp = / id="(.*?)"/;
+        const id = idRegexp.exec(msg)[1];
+        if (msg.indexOf('type="turn"') > -1) {
+          send(`<iq xmlns="jabber:client" type="result" to="${MOCK_USER.chat.jabberId}" id="${id}"><services xmlns="urn:xmpp:extdisco:1"><service transport="udp" port="3456" type="turn" username="turnuser:12395" password="akskdfjka=" host="turn.us-east-1.mypurecloud.com"/></services></iq>`);
+        } else {
+          send(`<iq xmlns="jabber:client" type="result" to="${MOCK_USER.chat.jabberId}" id="${id}"><services xmlns="urn:xmpp:extdisco:1"><service transport="udp" port="3456" type="stun" host="turn.us-east-1.mypurecloud.com"/></services></iq>`);
+        }
+      }
+    });
+  });
+  global.window.WebSocket = WebSocket;
+  ws = websocket;
+
+  return { getOrg, getUser, getChannel, getConversation, sendLogs, patchConversation, sdk, websocket };
+}
+
+module.exports = {
+  mockApis,
+  wss,
+  ws,
+  random,
+  timeout,
+  PARTICIPANT_ID
+};
