+++ test/unit/client.test.js
@@ -2,45 +2,11 @@
 
 const test = require('ava');
 const sinon = require('sinon');
-const nock = require('nock');
 const WildEmitter = require('wildemitter');
-const WebSocket = require('ws');
+
+let { wss, ws, mockApis, random, PARTICIPANT_ID } = require('../test-utils');
 
 const PureCloudWebrtcSdk = require('../../src/client');
-
-function random () {
-  return `${Math.random()}`.split('.')[1];
-}
-
-const USER_ID = random();
-const PARTICIPANT_ID = random();
-
-const MOCK_USER = {
-  id: USER_ID,
-  chat: { jabberId: 'hubert.j.farnsworth@planetexpress.mypurecloud.com' }
-};
-
-const MOCK_ORG = {
-  thirdPartyOrgId: '3000'
-};
-
-const MOCK_CONVERSATION = {
-  participants: [
-    {
-      id: PARTICIPANT_ID,
-      user: {
-        id: USER_ID
-      }
-    },
-    {
-      id: random()
-    }
-  ]
-};
-
-function timeout (n) {
-  return new Promise(resolve => setTimeout(resolve, n));
-}
 
 test.serial('constructor | throws if options are not provided', t => {
   t.throws(() => {
@@ -115,10 +81,10 @@ test.serial('constructor | sets up options when provided', t => {
   t.is(sdk._iceTransportPolicy, 'relay');
 });
 
-let wss;
-let ws;
 test.after(() => {
-  wss.close();
+  if (wss) {
+    wss.close();
+  }
 });
 
 const sandbox = sinon.createSandbox();
@@ -132,151 +98,6 @@ test.afterEach(() => {
   }
   sandbox.restore();
 });
-
-function mockApis ({ failOrg, failUser, failStreaming, failLogs, failLogsPayload, withMedia, conversationId, participantId, withLogs } = {}) {
-  nock.cleanAll();
-  const api = nock('https://api.mypurecloud.com');
-
-  // easy to debug nock
-  // api.log(console.error);
-
-  let getOrg;
-  if (failOrg) {
-    getOrg = api.get('/api/v2/organizations/me').reply(401);
-  } else {
-    getOrg = api.get('/api/v2/organizations/me').reply(200, MOCK_ORG);
-  }
-
-  let getUser;
-  if (failUser) {
-    getUser = api.get('/api/v2/users/me').reply(401);
-  } else {
-    getUser = api.get('/api/v2/users/me').reply(200, MOCK_USER);
-  }
-
-  const getChannel = api
-    .post('/api/v2/notifications/channels?connectionType=streaming')
-    .reply(200, { id: 'somechannelid' });
-
-  const conversationsApi = nock('https://api.mypurecloud.com');
-  let getConversation;
-  if (conversationId) {
-    getConversation = conversationsApi
-      .get(`/api/v2/conversations/calls/${conversationId}`)
-      .reply(200, MOCK_CONVERSATION);
-  }
-
-  let patchConversation;
-  if (conversationId && participantId) {
-    patchConversation = conversationsApi
-      .patch(`/api/v2/conversations/calls/${conversationId}/participants/${participantId}`)
-      .reply(202, {});
-  }
-
-  global.window = global.window || {};
-  if (withMedia) {
-    window.navigator = window.navigator || {};
-    window.navigator.mediaDevices = window.navigator.mediaDevices || {};
-    window.navigator.mediaDevices.getUserMedia = () => Promise.resolve(withMedia);
-  }
-
-  global.document = {
-    createElement: sinon.stub().returns({
-      addEventListener: (evt, callback) => setTimeout(callback, 10),
-      classList: { add () {} }
-    }),
-    querySelector () {},
-    body: {
-      append () {}
-    },
-    head: {
-      appendChild: sinon.stub().callsFake((script) => {
-        global.window = global.window || {};
-      })
-    }
-  };
-
-  const sdk = new PureCloudWebrtcSdk({
-    accessToken: '1234',
-    wsHost: failStreaming ? null : 'ws://localhost:1234',
-    logger: { debug () {}, log () {}, info () {}, warn () {}, error () {} }
-    // logger: { debug () {}, log () {}, info () {}, warn: console.warn.bind(console), error: console.error.bind(console) }
-  });
-
-  let sendLogs;
-  if (withLogs) {
-    const logsApi = nock('https://api.mypurecloud.com').persist();
-
-    if (failLogsPayload) {
-      sendLogs = logsApi.post('/api/v2/diagnostics/trace').replyWithError({status: 413, message: 'test fail'});
-    } else if (failLogs) {
-      sendLogs = logsApi.post('/api/v2/diagnostics/trace').replyWithError({status: 419, message: 'test fail'});
-    } else {
-      sendLogs = logsApi.post('/api/v2/diagnostics/trace').reply(200);
-    }
-  } else {
-    sdk._optOutOfTelemetry = true;
-  }
-
-  if (wss) {
-    wss.close();
-    wss = null;
-  }
-
-  wss = new WebSocket.Server({
-    port: 1234
-  });
-  let websocket;
-  wss.on('connection', function (ws) {
-    websocket = ws;
-    ws.on('message', function (msg) {
-      // console.error('⬆️', msg);
-      const send = function (r) {
-        // console.error('⬇️', r);
-        setTimeout(function () { ws.send(r); }, 15);
-      };
-      if (msg.indexOf('<open') === 0) {
-        send('<open xmlns="urn:ietf:params:xml:ns:xmpp-framing" xmlns:stream="http://etherx.jabber.org/streams" version="1.0" from="hawk" id="d6f681a3-358c-49df-819f-b231adb3cb97" xml:lang="en"></open>');
-        if (ws.__authSent) {
-          send(`<stream:features xmlns:stream="http://etherx.jabber.org/streams"><bind xmlns="urn:ietf:params:xml:ns:xmpp-bind"></bind><session xmlns="urn:ietf:params:xml:ns:xmpp-session"></session></stream:features>`);
-        } else {
-          send('<stream:features xmlns:stream="http://etherx.jabber.org/streams"><mechanisms xmlns="urn:ietf:params:xml:ns:xmpp-sasl"><mechanism>PLAIN</mechanism></mechanisms></stream:features>');
-        }
-      } else if (msg.indexOf('<auth') === 0) {
-        if (failStreaming) {
-          send('<failure xmlns="urn:ietf:params:xml:ns:xmpp-sasl"></failure>');
-        } else {
-          send('<success xmlns="urn:ietf:params:xml:ns:xmpp-sasl"></success>');
-        }
-        ws.__authSent = true;
-      } else if (msg.indexOf('<bind') !== -1) {
-        const idRegexp = /id="(.*?)"/;
-        const id = idRegexp.exec(msg)[1];
-        send(`<iq xmlns="jabber:client" id="${id}" to="${MOCK_USER.chat.jabberId}" type="result"><bind xmlns="urn:ietf:params:xml:ns:xmpp-bind"><jid>${MOCK_USER.chat.jabberId}/d6f681a3-358c-49df-819f-b231adb3cb97</jid></bind></iq>`);
-      } else if (msg.indexOf('<session') !== -1) {
-        const idRegexp = /id="(.*?)"/;
-        const id = idRegexp.exec(msg)[1];
-        send(`<iq xmlns="jabber:client" id="${id}" to="${MOCK_USER.chat.jabberId}/d6f681a3-358c-49df-819f-b231adb3cb97" type="result"></iq>`);
-      } else if (msg.indexOf('ping') !== -1) {
-        const idRegexp = /id="(.*?)"/;
-        const id = idRegexp.exec(msg)[1];
-        send(`<iq xmlns="jabber:client" to="${MOCK_USER.chat.jabberId}" type="result" id="${id}"></iq>`);
-      } else if (msg.indexOf('extdisco') !== -1) {
-        const idRegexp = / id="(.*?)"/;
-        const id = idRegexp.exec(msg)[1];
-        if (msg.indexOf('type="turn"') > -1) {
-          send(`<iq xmlns="jabber:client" type="result" to="${MOCK_USER.chat.jabberId}" id="${id}"><services xmlns="urn:xmpp:extdisco:1"><service transport="udp" port="3456" type="turn" username="turnuser:12395" password="akskdfjka=" host="turn.us-east-1.mypurecloud.com"/></services></iq>`);
-        } else {
-          send(`<iq xmlns="jabber:client" type="result" to="${MOCK_USER.chat.jabberId}" id="${id}"><services xmlns="urn:xmpp:extdisco:1"><service transport="udp" port="3456" type="stun" host="turn.us-east-1.mypurecloud.com"/></services></iq>`);
-        }
-      }
-    });
-  });
-  global.window.WebSocket = WebSocket;
-  ws = websocket;
-
-  return { getOrg, getUser, getChannel, getConversation, sendLogs, patchConversation, sdk, websocket };
-}
 
 test.serial('initialize | fetches org and person details, sets up the streaming connection', async t => {
   const { getOrg, getUser, getChannel, sdk } = mockApis();
@@ -733,293 +554,6 @@ test.serial('onSession | uses existing media, attaches it to the session, attach
   sinon.assert.notCalled(global.document.body.append);
 
   sandbox.restore();
-});
-
-test.serial('_log | will not notify logs if the logLevel is lower than configured', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  sdk._logLevel = 'warn';
-  await sdk.initialize();
-  sinon.stub(sdk, '_notifyLogs');
-  sdk._log('debug', 'test', { details: 'etc' });
-  sinon.assert.notCalled(sdk._notifyLogs);
-});
-
-test.serial('_log | will not notify logs if opted out', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  sdk._logLevel = 'debug';
-  sdk._optOutOfTelemetry = true;
-  await sdk.initialize();
-  sinon.stub(sdk, '_notifyLogs');
-  sdk._log('warn', 'test', { details: 'etc' });
-  sinon.assert.notCalled(sdk._notifyLogs);
-});
-
-test.serial('_log | will buffer a log and notify it if the logLevel is gte configured', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  sdk._logLevel = 'warn';
-  await sdk.initialize();
-  sinon.stub(sdk, '_notifyLogs');
-  console.log(sdk._logBuffer[0]);
-  t.is(sdk._logBuffer.length, 0);
-  sdk._log('warn', 'test', { details: 'etc' });
-  sinon.assert.calledOnce(sdk._notifyLogs);
-  t.is(sdk._logBuffer.length, 1);
-});
-
-test.serial('_notifyLogs | will debounce logs and only send logs once at the end', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  sdk._logLevel = 'warn';
-  await sdk.initialize();
-  sinon.stub(sdk, '_sendLogs');
-  t.is(sdk._logBuffer.length, 0);
-  sdk._log('warn', 'test', { details: 'etc' });
-  sinon.assert.notCalled(sdk._sendLogs);
-  for (let i = 1; i < 6; i++) {
-    await timeout(100 * i);
-    sdk._log('warn', 'test' + i);
-  }
-  sinon.assert.notCalled(sdk._sendLogs);
-  t.is(sdk._logBuffer.length, 6);
-  await timeout(1100);
-  sinon.assert.calledOnce(sdk._sendLogs);
-});
-
-test.serial('_sendLogs | resets all flags related to backoff on success', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  sdk._logLevel = 'warn';
-  await sdk.initialize();
-
-  sdk._backoffActive = true;
-  sdk._failedLogAttempts = 2;
-  sdk._reduceLogPayload = true;
-  sdk._logBuffer.push('log1');
-
-  await sdk._sendLogs();
-  t.is(sdk._backoffActive, false);
-  t.is(sdk._failedLogAttempts, 0);
-  t.is(sdk._reduceLogPayload, false);
-});
-
-test.serial('_sendLogs | resets the backoff on success', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  sdk._logLevel = 'warn';
-  await sdk.initialize();
-
-  const backoffResetSpy = sinon.spy(sdk._backoff, 'reset');
-  sdk._logBuffer.push('log1');
-  sdk._logBuffer.push('log1');
-
-  await sdk._sendLogs();
-  sinon.assert.calledOnce(backoffResetSpy);
-});
-
-test.serial('_sendLogs | should call backoff.backoff() again if there are still items in the _logBuffer after a successfull call to api', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  sdk._logLevel = 'warn';
-  await sdk.initialize();
-
-  const backoffSpy = sinon.spy(sdk._backoff, 'backoff');
-  sdk._reduceLogPayload = true;
-  sdk._logBuffer.push('log1');
-  sdk._logBuffer.push('log2');
-  sdk._logBuffer.push('log3');
-  sdk._logBuffer.push('log4');
-
-  await sdk._sendLogs();
-  sinon.assert.calledOnce(backoffSpy);
-});
-
-test.serial('_sendLogs | will add logs back to buffer if request fails', async t => {
-  const expectedFirstLog = 'log1';
-  const expectedSecondLog = 'log2';
-  const expectedThirdLog = 'log3';
-  let { sdk } = mockApis({failLogs: true, withLogs: true});
-  sdk._logLevel = 'warn';
-  await sdk.initialize();
-
-  t.is(sdk._logBuffer.length, 0);
-  sdk._logBuffer.push(expectedFirstLog);
-  sdk._logBuffer.push(expectedSecondLog);
-  sdk._logBuffer.push(expectedThirdLog);
-
-  await sdk._sendLogs();
-
-  t.is(sdk._logBuffer.length, 3);
-  t.is(sdk._logBuffer[0], expectedFirstLog, 'Log items should be put back into the buffer the same way they went out');
-  t.is(sdk._logBuffer[1], expectedSecondLog, 'Log items should be put back into the buffer the same way they went out');
-  t.is(sdk._logBuffer[2], expectedThirdLog, 'Log items should be put back into the buffer the same way they went out');
-  sdk.logBuffer = [];
-  sdk._optOutOfTelemetry = true;
-});
-
-test.serial('_sendLogs | increments _failedLogAttemps on failure', async t => {
-  const { sdk } = mockApis({failLogsPayload: true, withLogs: true});
-  sdk._logLevel = 'warn';
-  await sdk.initialize();
-  t.is(sdk._logBuffer.length, 0);
-  sdk._logBuffer.push('log1');
-  sdk._logBuffer.push('log2');
-  t.is(sdk._failedLogAttempts, 0);
-
-  await sdk._sendLogs();
-  t.is(sdk._failedLogAttempts, 1);
-});
-
-test.serial('_sendLogs | sets _reduceLogPayload to true if error status is 413 (payload too large)', async t => {
-  const { sdk } = mockApis({failLogsPayload: true, withLogs: true});
-  sdk._logLevel = 'warn';
-  await sdk.initialize();
-  t.is(sdk._logBuffer.length, 0);
-  sdk._logBuffer.push('log1');
-  sdk._logBuffer.push('log2');
-  t.is(sdk._reduceLogPayload, false);
-
-  await sdk._sendLogs();
-  t.is(sdk._reduceLogPayload, true);
-});
-
-test.serial('_sendLogs | should reset all backoff flags and reset the backoff if api request returns error and payload was only 1 log', async t => {
-  const { sdk } = mockApis({failLogsPayload: true, withLogs: true});
-  sdk._logLevel = 'warn';
-  await sdk.initialize();
-  sdk._logBuffer.push('log1');
-  const backoffResetSpy = sinon.spy(sdk._backoff, 'reset');
-
-  await sdk._sendLogs();
-  t.is(sdk._backoffActive, false);
-  t.is(sdk._failedLogAttempts, 0);
-  t.is(sdk._reduceLogPayload, false);
-  sinon.assert.calledOnce(backoffResetSpy);
-});
-
-test.serial('_sendLogs | set backoffActive to false if the backoff fails', async t => {
-  const { sdk } = mockApis({failLogs: true, withLogs: true});
-  sdk._logLevel = 'warn';
-  sinon.spy(sdk, '_sendLogs');
-  await sdk.initialize();
-  sdk._log('error', 'log1');
-  sdk._log('error', 'log2');
-
-  sdk._backoff.failAfter(1); // means it will retry once, or 2 tries total
-
-  sdk._notifyLogs();
-  sdk._notifyLogs();
-  await timeout(1000);
-  sdk._notifyLogs();
-  await timeout(5000);
-  sinon.assert.calledTwice(sdk._sendLogs);
-  t.is(sdk._backoffActive, false);
-});
-
-test.serial('_getLogPayload | returns the entire _logBuffer if _reduceLogPayload is false', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  await sdk.initialize();
-  sdk._reduceLogPayload = false;
-  sdk._logBuffer = [0, 1, 2, 3, 4];
-
-  const result = sdk._getLogPayload();
-
-  t.is(result.length, 5);
-  t.is(result[0], 0);
-  t.is(result[1], 1);
-  t.is(result[2], 2);
-  t.is(result[3], 3);
-  t.is(result[4], 4);
-  t.is(sdk._logBuffer.length, 0, 'Items should have been removed from _logBuffer');
-});
-
-test.serial('_getLogPayload | returns part of _logBuffer if _reduceLogPayload is true', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  await sdk.initialize();
-  sdk._reduceLogPayload = true;
-  sdk._failedLogAttempts = 1;
-  sdk._logBuffer = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
-
-  const result = sdk._getLogPayload();
-
-  t.is(result.length, 5);
-  t.is(result[0], 0);
-  t.is(result[1], 1);
-  t.is(result[2], 2);
-  t.is(result[3], 3);
-  t.is(result[4], 4);
-
-  t.is(sdk._logBuffer.length, 5, 'Items should have been removed from _logBuffer');
-  t.is(sdk._logBuffer[0], 5);
-  t.is(sdk._logBuffer[1], 6);
-  t.is(sdk._logBuffer[2], 7);
-  t.is(sdk._logBuffer[3], 8);
-  t.is(sdk._logBuffer[4], 9);
-});
-
-test.serial('_getLogPayload | returns part of _logBuffer if _reduceLogPayload is true and _failedLogAttempts is 0', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  await sdk.initialize();
-  sdk._reduceLogPayload = true;
-  sdk._failedLogAttempts = 0;
-  sdk._logBuffer = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
-
-  const result = sdk._getLogPayload();
-
-  t.is(result.length, 5);
-  t.is(result[0], 0);
-  t.is(result[1], 1);
-  t.is(result[2], 2);
-  t.is(result[3], 3);
-  t.is(result[4], 4);
-
-  t.is(sdk._logBuffer.length, 5, 'Items should have been removed from _logBuffer');
-  t.is(sdk._logBuffer[0], 5);
-  t.is(sdk._logBuffer[1], 6);
-  t.is(sdk._logBuffer[2], 7);
-  t.is(sdk._logBuffer[3], 8);
-  t.is(sdk._logBuffer[4], 9);
-});
-
-test.serial('_resetBackoffFlags | should reset values of _backoffActive, _failedLogAttempts, and _reduceLogPaylod', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  await sdk.initialize();
-  sdk._backoffActive = true;
-  sdk._failedLogAttempts = 3;
-  sdk._reduceLogPayload = true;
-
-  sdk._resetBackoffFlags();
-
-  t.is(sdk._backoffActive, false);
-  t.is(sdk._failedLogAttempts, 0);
-  t.is(sdk._reduceLogPayload, false);
-});
-
-test.serial('_getReducedLogPayload | should return at least one log item', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  await sdk.initialize();
-
-  sdk._logBuffer = [1, 2, 3, 4, 5];
-  const result = sdk._getReducedLogPayload(6);
-
-  t.is(result.length, 1);
-});
-
-test.serial('_getReducesLogPayload | should remove items from _logBuffer and return them', async t => {
-  const { sdk } = mockApis({ withLogs: true });
-  await sdk.initialize();
-
-  sdk._logBuffer = [0, 1, 2, 3, 4, 5, 6, 7, 8, 9];
-  const result = sdk._getReducedLogPayload(1);
-
-  t.is(result.length, 5);
-  t.is(result[0], 0);
-  t.is(result[1], 1);
-  t.is(result[2], 2);
-  t.is(result[3], 3);
-  t.is(result[4], 4);
-
-  t.is(sdk._logBuffer.length, 5, 'Items should have been removed from the _logBuffer');
-  t.is(sdk._logBuffer[0], 5);
-  t.is(sdk._logBuffer[1], 6);
-  t.is(sdk._logBuffer[2], 7);
-  t.is(sdk._logBuffer[3], 8);
-  t.is(sdk._logBuffer[4], 9);
 });
 
 test.serial('_refreshTurnServers | refreshes the turn servers', async t => {
